docs(list): update demo
